+++ src/main/java/com/fox/spider/stock/entity/po/tencent/TencentRealtimeDealInfoPo.java
@@ -102,11 +102,11 @@ public class TencentRealtimeDealInfoPo implements Serializable {
     /**
      * 排名靠前的5个买方报价
      */
-    LinkedHashMap<BigDecimal, Integer> buyPriceMap;
+    LinkedHashMap<BigDecimal, Long> buyPriceMap;
     /**
      * 排名靠前的5个卖方报价
      */
-    LinkedHashMap<BigDecimal, Integer> sellPriceMap;
+    LinkedHashMap<BigDecimal, Long> sellPriceMap;
     /**
      * 总市值
      */
